backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
google-cloud-storage==2.19.0
supabase==2.11.0
pydantic==2.10.5
pydantic-settings==2.7.1
httpx==0.28.1
